examples/pil_example/additional_requirements.txt
Pillow==6.0.0
